@@ -10,13 +10,13 @@ dependencies = [
     "numpy>=1.18.0",
     "scipy>=1.4.0",
     "scikit-learn>=0.22.0",
-#    "sentence-transformers>=0.3.0",
     "pymilvus>=2.0.0",
     "pydantic>=1.8.0",
     "boto3>=1.17.0",
     "pyarrow>=16.1.0",
+    "fastembed>=0.3.4",
 ]
-requires-python = ">=3.10"
+requires-python = "<=3.12,>=3.10"
 readme = "README.md"
 license = {text = "MIT"}
 
